src/requirements.txt
asyncpg
discord.py>=1.2.5
psutil
pillow
fuzzywuzzy[speedup]